fix: files with structure
