+++ src/app/members/actions/createUser.ts
@@ -1,12 +1,12 @@
 "use server";
 import { Tables } from "@/lib/db";
-import { getSupabaseClient } from "../../../../utils/supabase/supabaseClient";
+import { getSupabaseClient } from "../../../utils/supabase/supabaseClient";
 import { UserStatus } from "@/constant/constant";
 import { redirect } from "next/navigation";
-import { Routes } from "@/lib/constants";
-import { formatPhoneNumber } from "../../../../utils/formatPhoneNumber";
+import { formatPhoneNumber } from "../../../utils/formatPhoneNumber";
 import { UserBuilder } from "../components/MemberFormBuilder";
 import { headers } from "next/headers";
+import { Routes } from "@/lib/routes";
 
 export const createUser = async (values: UserBuilder) => {
   const supabase = await getSupabaseClient();
