+++ src/app/members/actions/logoutUser.ts
@@ -1,5 +1,5 @@
 import { Routes } from "@/lib/routes";
-import { clearCookies } from "../../../../utils/cookiesManager";
+import { clearCookies } from "../../../utils/cookiesManager";
 import { redirect } from "next/navigation";
 
 export const logoutUser = async () => {
