+++ src/app/members/components/MemberActionRender.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ActionsMenu from "@/common/ActionMenu/ActionsMenu";
 import { Info, Trash2, Edit, Send } from "lucide-react";
-import { MemberRole } from "@/lib/constants";
+import { MemberRole } from "@/constant/constant";
 
 interface MemberInfo {
   id: number;
