+++ src/app/members/components/MemberDetails.tsx
@@ -10,7 +10,7 @@ import {
   User as UserIcon,
 } from "lucide-react";
 import { UserDetails } from "@/types";
-import { MemberRole } from "@/lib/constants";
+import { MemberRole } from "@/constant/constant";
 
 export const MemberDetails = ({ userInfo }: { userInfo: UserDetails }) => {
   return (
