+++ src/app/members/components/MemberDetials.tsx
@@ -9,7 +9,7 @@ import {
   Shield as ShieldIcon,
   User as UserIcon } from "lucide-react";
 import { UserDetails } from "@/types";
-import { MemberRole } from "@/lib/constants";
+import { MemberRole } from "@/constant/constant";
 
 export const MemberDetials = ({
     userInfo
