+++ src/constant/constant.ts
@@ -1,3 +1,10 @@
+export enum MemberRole {
+  Incharge = "incharge",
+  ShiftIncharge = "shift-incharge",
+  Member = "member",
+  SuperAdmin = "super_admin",
+}
+
 export enum UserRole {
   Member = "member",
   Incharge = "incharge",
@@ -24,4 +31,16 @@ export enum LeaveTypes {
   Sick = "sick",
   Vacation = "vacation",
   Personal = "personal",
+}
+
+export enum AttendanceStatus {
+  Pending = "pending",
+  Accept = "accept",
+  Reject = "reject",
+}
+
+export enum Table {
+  Attendance = "attendance",
+  Member = "member",
+  User = "user",
 }
