+++ src/lib/routes.tsx
@@ -13,21 +13,3 @@ export enum Routes {
   LeaveRequest = "/leaves",
   AddLeaveRequest = "/leaves/add",
 }
-
-export enum MemberRole {
-  Incharge = "incharge",
-  ShiftIncharge = "shift-incharge",
-  Member = "member",
-  SuperAdmin = "super_admin",
-}
-
-export enum UserRole {
-  Incharge = "incharge",
-  Member = "member",
-  ShiftIncharge = "shift-incharge",
-}
-
-export enum UserStatus {
-  Active = "active",
-  Invited = "invited",
-}
