+++ src/utils/supabase/check-env-vars.ts
