+++ tsconfig.json
@@ -22,7 +22,7 @@
     "paths": {
       "@/*": ["./src/*"],
       "@/lib/*": ["./src/lib/*"],
-      "@/utils/*" : ["./utils/*"]
+      "@/utils/*" : ["src/utils/*"]
       
     }
